address comments: rename tracing to io_stats
