+++ components/file_system/src/lib.rs
@@ -13,18 +13,18 @@ extern crate test;
 extern crate tikv_alloc;
 
 mod file;
+mod io_stats;
 mod metrics;
 mod metrics_manager;
 mod rate_limiter;
-mod tracing;
 
 pub use file::{File, OpenOptions};
+pub use io_stats::{get_io_type, init as init_io_stats_collector, set_io_type};
 pub use metrics_manager::{BytesFetcher, MetricsManager};
 pub use rate_limiter::{
     get_io_rate_limiter, set_io_rate_limiter, IOBudgetAdjustor, IORateLimitMode, IORateLimiter,
     IORateLimiterStatistics,
 };
-pub use tracing::{get_io_type, init as init_tracing, set_io_type};
 
 pub use std::fs::{
     canonicalize, create_dir, create_dir_all, hard_link, metadata, read_dir, read_link, remove_dir,
